Fix docs for LocalSnapshot
* Moved everything into data namespace just to have a unified location
for it. I removed hoomd/data.py which wasn't being used as far as I was
aware.
* I had to document the individual property classes for showing
availalbe particle, bond, ... properties.
