+++ hoomd/__init__.py
@@ -39,7 +39,6 @@ from hoomd import analyze
 from hoomd import benchmark
 from hoomd import comm
 from hoomd import compute
-from hoomd import data
 from hoomd import dump
 from hoomd import group
 from hoomd import init
