+++ hoomd/data/__init__.py
@@ -1,0 +1,5 @@
+from .local_access import (
+    LocalSnapshot, AngleLocalAccessCPU, BondLocalAccessCPU,
+    ConstraintLocalAccessCPU, DihedralLocalAccessCPU, ImproperLocalAccessCPU,
+    PairLocalAccessCPU, ParticleLocalAccessCPU)
+from .local_access_gpu import LocalSnapshotGPU
