+++ sphinx-doc/module-hoomd-data.rst
@@ -1,0 +1,36 @@
+hoomd.data
+------------------
+
+.. rubric:: Overview
+
+.. py:currentmodule:: hoomd.data
+
+.. autosummary::
+    :nosignatures:
+
+    LocalSnapshot
+    LocalSnapshotGPU
+    AngleLocalAccessCPU
+    BondLocalAccessCPU
+    ConstraintLocalAccessCPU
+    DihedralLocalAccessCPU
+    ImproperLocalAccessCPU
+    PairLocalAccessCPU
+    ParticleLocalAccessCPU
+
+.. rubric:: Details
+
+.. automodule:: hoomd.data
+    :synopsis: Provide access in Python to data buffers on CPU.
+    :members: AngleLocalAccessCPU,
+              BondLocalAccessCPU,
+              ConstraintLocalAccessCPU,
+              DihedralLocalAccessCPU,
+              ImproperLocalAccessCPU,
+              PairLocalAccessCPU,
+              ParticleLocalAccessCPU
+    
+    .. autoclass:: LocalSnapshot
+        :inherited-members:
+
+    .. autoclass:: LocalSnapshotGPU
