--- sphinx-doc/module-hoomd-local-gpu.rst
@@ -1,18 +1,0 @@
-hoomd.local_access_gpu
-----------------------
-
-.. rubric:: Overview
-
-.. autosummary::
-    :nosignatures:
-
-    hoomd.local_access_gpu.LocalSnapshotGPU
-
-.. rubric:: Details
-
-.. automodule:: hoomd.local_access
-    :synopsis: Provide access in Python to data buffers on GPU.
-    :undoc-members:
-
-    .. autoclass:: hoomd.local_access_gpu.LocalSnapshotGPU
-        :members: box, local_box, particles, bonds, angles, dihedrals, impropers, constraints, pairs
