--- sphinx-doc/module-hoomd-local.rst
@@ -1,18 +1,0 @@
-hoomd.local_access
-------------------
-
-.. rubric:: Overview
-
-.. autosummary::
-    :nosignatures:
-
-    hoomd.local_access.LocalSnapshot
-
-.. rubric:: Details
-
-.. automodule:: hoomd.local_access
-    :synopsis: Provide access in Python to data buffers on CPU.
-    :undoc-members:
-
-    .. autoclass:: hoomd.local_access.LocalSnapshot
-        :members: box, local_box, particles, bonds, angles, dihedrals, impropers, constraints, pairs
