+++ sphinx-doc/package-hoomd.rst
@@ -37,12 +37,11 @@ hoomd
    module-hoomd-comm
    module-hoomd-compute
    module-hoomd-custom
+   module-hoomd-data
    module-hoomd-device
    module-hoomd-dump
    module-hoomd-filter
    module-hoomd-integrate
-   module-hoomd-local
-   module-hoomd-local-gpu
    module-hoomd-logging
    module-hoomd-triggers
    module-hoomd-update
